Develop

+++ server/controllers/portfolio.js
@@ -55,8 +55,10 @@ const getportfolio = async (req, res) => {
 
 const deletePortfolio = async (req, res) => {
   const portfolioSlug = req.params.portfolioSlug;
-  const portfolio = await Portfolio.findOneAndDelete({ portfolioSlug });
-  res.json(portfolio);
+  const portfolio = await Portfolio.findOne({ portfolioSlug });
+  if (!portfolio) { return res.status(400).send("Portfolio Not Found"); }
+  const result = await Portfolio.findOneAndDelete({ portfolioSlug });
+  res.json({success:true});
 }
 
 // Get a particular Module
@@ -72,8 +74,10 @@ const getmodule = async (req, res) => {
 const deletemodule = async (req, res) => {
   const portfolioSlug = req.params.portfolioSlug;
   const i = req.params.moduleNumber;
-  const portfolio = await Portfolio.updateOne({ portfolioSlug }, { $pull: { modules: { moduleNumber: i } } });
-  res.json(portfolio);
+  const portfolio = await Portfolio.findOne({ portfolioSlug });
+  if (!portfolio) { return res.status(400).send("Portfolio Not Found"); }
+  portfolio = await Portfolio.updateOne({ portfolioSlug }, { $pull: { modules: { moduleNumber: i } } });
+  res.json({portfolio,success:true});
 }
 
 // Add a new module
@@ -168,13 +172,7 @@ const addVideo = async (req, res) => {
         videoSlug,
         videoLink: data.Location
       };
-      //    if (videoCount === videoNumber - 1) {
-      //   const NEW_DOC = {
-      //     videoTitle,
-      //     videoNumber,
-      //     videoSlug,
-      //   };
-
+    
       const result = await Portfolio.findOneAndUpdate(
         {
           portfolioSlug,
@@ -189,85 +187,12 @@ const addVideo = async (req, res) => {
         }
       );
       return res.status(200).json(result);
-      // } else if (videoCount !== i - 1) {
-      //   return res
-      //     .status(400)
-      //     .send(
-      //       "Video Number Alreay Exists or Please enter the latest Video Number"
-      //     );
-      // } else {
-      //   return res.status(400).send("Video Number Not Exists");
-      // }
-      // console.log(data);
-      // return res.status(200).json(data);
     });
-
-
-    // res.status(200).json(video);
   } catch (err) {
     console.log(err);
     return res.status(400).send({ err });
   }
 };
-
-// const addVideo = async (req, res) => {
-//   const { videoTitle, videoNumber } = req.body;
-//   const videoSlug = slugify(req.body.videoTitle, "_");
-//   const useremail = req.email;
-//   try {
-//     const portfolioSlug = req.params.portfolioSlug;
-//     const i = req.params.moduleNumber;
-
-//     const portfolio = await Portfolio.findOne({ portfolioSlug });
-//     if (!portfolio) {
-//       return res.status(400).send("Portfolio Not Found");
-//     }
-//     if (useremail != portfolio.portfolioCreator) {
-//       return res.status(400).send("Unauthorized");
-//     }
-
-//     const Modulescount = Object.keys(portfolio.modules).length;
-//     if (i < 1 || i > Modulescount) {
-//       return res.status(400).send("Enter a valid module number");
-//     }
-
-//     const videoCount = Object.keys(portfolio.modules[i - 1].videos).length;
-
-//     if (videoCount === videoNumber - 1) {
-//       const NEW_DOC = {
-//         videoTitle,
-//         videoNumber,
-//         videoSlug,
-//       };
-
-//       const result = await Portfolio.findOneAndUpdate(
-//         {
-//           portfolioSlug,
-//           modules: {
-//             $elemMatch: {
-//               moduleNumber: i,
-//             },
-//           },
-//         },
-//         {
-//           $push: { "modules.$.videos": NEW_DOC },
-//         }
-//       );
-//       res.status(200).json({ result: result });
-//     } else if (videoCount !== i - 1) {
-//       return res
-//         .status(400)
-//         .send(
-//           "Video Number Alreay Exists or Please enter the latest Video Number"
-//         );
-//     } else {
-//       return res.status(400).send("Video Number Not Exists");
-//     }
-//   } catch (err) {
-//     console.log(err);
-//     return res.status(400).send({ err });
-//   }
-// };
 
 module.exports = {
   createPortfolio,
